Extract site content fragment and tests
