+++ WordPress/src/main/java/org/wordpress/android/ui/mysite/MySiteContentFragment.kt
@@ -1,0 +1,310 @@
+package org.wordpress.android.ui.mysite
+
+import android.app.Activity
+import android.content.Intent
+import android.os.Bundle
+import android.os.Parcelable
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.fragment.app.Fragment
+import androidx.lifecycle.ViewModelProvider
+import androidx.recyclerview.widget.LinearLayoutManager
+import com.google.android.material.snackbar.Snackbar
+import com.yalantis.ucrop.UCrop
+import kotlinx.android.synthetic.main.my_site_content_fragment.*
+import org.wordpress.android.R
+import org.wordpress.android.WordPress
+import org.wordpress.android.fluxc.model.SiteModel
+import org.wordpress.android.ui.RequestCodes
+import org.wordpress.android.ui.TextInputDialogFragment
+import org.wordpress.android.ui.domains.DomainRegistrationResultFragment.Companion.RESULT_REGISTERED_DOMAIN_EMAIL
+import org.wordpress.android.ui.main.WPMainActivity
+import org.wordpress.android.ui.mysite.SiteIconUploadHandler.ItemUploadedModel
+import org.wordpress.android.ui.pages.SnackbarMessageHolder
+import org.wordpress.android.ui.photopicker.MediaPickerConstants
+import org.wordpress.android.ui.photopicker.MediaPickerLauncher
+import org.wordpress.android.ui.photopicker.PhotoPickerActivity.PhotoPickerMediaSource
+import org.wordpress.android.ui.posts.BasicDialogViewModel
+import org.wordpress.android.ui.posts.BasicDialogViewModel.BasicDialogModel
+import org.wordpress.android.ui.uploads.UploadService
+import org.wordpress.android.ui.uploads.UploadUtilsWrapper
+import org.wordpress.android.ui.utils.UiHelpers
+import org.wordpress.android.util.AppLog
+import org.wordpress.android.util.AppLog.T.MAIN
+import org.wordpress.android.util.AppLog.T.UTILS
+import org.wordpress.android.util.SnackbarItem
+import org.wordpress.android.util.SnackbarItem.Action
+import org.wordpress.android.util.SnackbarItem.Info
+import org.wordpress.android.util.SnackbarSequencer
+import org.wordpress.android.util.ToastUtils.showToast
+import org.wordpress.android.util.image.ImageManager
+import org.wordpress.android.util.setVisible
+import javax.inject.Inject
+
+class MySiteContentFragment : Fragment(),
+        TextInputDialogFragment.Callback {
+    @Inject lateinit var viewModelFactory: ViewModelProvider.Factory
+    @Inject lateinit var imageManager: ImageManager
+    @Inject lateinit var uiHelpers: UiHelpers
+    @Inject lateinit var snackbarSequencer: SnackbarSequencer
+    @Inject lateinit var mediaPickerLauncher: MediaPickerLauncher
+    @Inject lateinit var uploadUtilsWrapper: UploadUtilsWrapper
+    @Inject lateinit var mySiteNavigationActionHandler: MySiteNavigationActionHandler
+    private lateinit var viewModel: MySiteContentViewModel
+    private lateinit var dialogViewModel: BasicDialogViewModel
+    private lateinit var quickStartMenuViewModel: QuickStartMenuViewModel
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        val siteId = savedInstanceState?.getInt(KEY_SITE_ID) ?: arguments?.getInt(KEY_SITE_ID)
+        ?: throw IllegalArgumentException("Site ID must be defined")
+        (requireActivity().application as WordPress).component().inject(this)
+        viewModel = ViewModelProvider(this, viewModelFactory).get(MySiteContentViewModel::class.java)
+        dialogViewModel = ViewModelProvider(requireActivity(), viewModelFactory)
+                .get(BasicDialogViewModel::class.java)
+        quickStartMenuViewModel = ViewModelProvider(requireActivity(), viewModelFactory)
+                .get(QuickStartMenuViewModel::class.java)
+        viewModel.start(siteId)
+    }
+
+    override fun onCreateView(
+        inflater: LayoutInflater,
+        container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        return inflater.inflate(
+                R.layout.my_site_content_fragment,
+                container,
+                false
+        )
+    }
+
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+
+        val layoutManager = LinearLayoutManager(activity)
+
+        savedInstanceState?.getParcelable<Parcelable>(KEY_LIST_STATE)?.let {
+            layoutManager.onRestoreInstanceState(it)
+        }
+
+        recycler_view.layoutManager = layoutManager
+
+        val adapter = MySiteAdapter(imageManager, uiHelpers)
+
+        savedInstanceState?.getBundle(KEY_NESTED_LISTS_STATES)?.let {
+            adapter.onRestoreInstanceState(it)
+        }
+
+        recycler_view.adapter = adapter
+
+        viewModel.uiModel.observe(viewLifecycleOwner, {
+            it?.let { uiModel ->
+                loadData(uiModel)
+            }
+        })
+        viewModel.onBasicDialogShown.observe(viewLifecycleOwner, {
+            it?.getContentIfNotHandled()?.let { model ->
+                dialogViewModel.showDialog(requireActivity().supportFragmentManager,
+                        BasicDialogModel(
+                                model.tag,
+                                getString(model.title),
+                                getString(model.message),
+                                getString(model.positiveButtonLabel),
+                                model.negativeButtonLabel?.let { label -> getString(label) },
+                                model.cancelButtonLabel?.let { label -> getString(label) }
+                        ))
+            }
+        })
+        viewModel.onTextInputDialogShown.observe(viewLifecycleOwner, {
+            it?.getContentIfNotHandled()?.let { model ->
+                val inputDialog = TextInputDialogFragment.newInstance(
+                        getString(model.title),
+                        model.initialText,
+                        getString(model.hint),
+                        model.isMultiline,
+                        model.isInputEnabled,
+                        model.callbackId
+                )
+                inputDialog.setTargetFragment(this, 0)
+                inputDialog.show(parentFragmentManager, TextInputDialogFragment.TAG)
+            }
+        })
+        viewModel.onQuickStartMenuShown.observe(viewLifecycleOwner, {
+            it?.getContentIfNotHandled()?.let { id ->
+                ((parentFragmentManager.findFragmentByTag(id) as? QuickStartMenuFragment)
+                        ?: QuickStartMenuFragment.newInstance(id))
+                        .show(parentFragmentManager, id)
+            }
+        })
+        viewModel.onNavigation.observe(viewLifecycleOwner, {
+            it?.getContentIfNotHandled()?.let { action ->
+                mySiteNavigationActionHandler.navigate(requireActivity(), this, action)
+            }
+        })
+        viewModel.onSnackbarMessage.observe(viewLifecycleOwner, {
+            it?.getContentIfNotHandled()?.let { messageHolder ->
+                showSnackbar(messageHolder)
+            }
+        })
+        viewModel.onMediaUpload.observe(viewLifecycleOwner, {
+            it?.getContentIfNotHandled()?.let { mediaModel ->
+                UploadService.uploadMedia(requireActivity(), mediaModel)
+            }
+        })
+        dialogViewModel.onInteraction.observe(viewLifecycleOwner, {
+            it?.getContentIfNotHandled()?.let { interaction -> viewModel.onDialogInteraction(interaction) }
+        })
+        quickStartMenuViewModel.onInteraction.observe(viewLifecycleOwner, {
+            // TODO Handle Quick Start menu interaction
+            it?.getContentIfNotHandled()?.let { interaction -> showToast(context, interaction.toString()) }
+        })
+        viewModel.onUploadedItem.observe(viewLifecycleOwner, {
+            it?.getContentIfNotHandled()?.let { itemUploadedModel ->
+                when (itemUploadedModel) {
+                    is ItemUploadedModel.PostUploaded -> {
+                        uploadUtilsWrapper.onPostUploadedSnackbarHandler(
+                                activity,
+                                requireActivity().findViewById(R.id.coordinator), true,
+                                itemUploadedModel.post, itemUploadedModel.errorMessage, itemUploadedModel.site
+                        )
+                    }
+                    is ItemUploadedModel.MediaUploaded -> {
+                        uploadUtilsWrapper.onMediaUploadedSnackbarHandler(
+                                activity,
+                                requireActivity().findViewById(R.id.coordinator), true,
+                                itemUploadedModel.media, itemUploadedModel.site, itemUploadedModel.errorMessage
+                        )
+                    }
+                }
+            }
+        })
+    }
+
+    override fun onResume() {
+        super.onResume()
+        viewModel.refresh()
+    }
+
+    override fun onSaveInstanceState(outState: Bundle) {
+        super.onSaveInstanceState(outState)
+        recycler_view.layoutManager?.let {
+            outState.putParcelable(KEY_LIST_STATE, it.onSaveInstanceState())
+        }
+        (recycler_view.adapter as? MySiteAdapter)?.let {
+            outState.putBundle(KEY_NESTED_LISTS_STATES, it.onSaveInstanceState())
+        }
+        outState.putInt(KEY_SITE_ID, viewModel.site.id)
+    }
+
+    override fun onActivityResult(requestCode: Int, resultCode: Int, data: Intent?) {
+        super.onActivityResult(requestCode, resultCode, data)
+        if (data == null) {
+            return
+        }
+        when (requestCode) {
+            RequestCodes.DO_LOGIN -> if (resultCode == Activity.RESULT_OK) {
+                viewModel.handleSuccessfulLoginResult()
+            }
+            RequestCodes.SITE_ICON_PICKER -> {
+                if (resultCode != Activity.RESULT_OK) {
+                    return
+                }
+                when {
+                    data.hasExtra(MediaPickerConstants.EXTRA_MEDIA_ID) -> {
+                        val mediaId = data.getLongExtra(MediaPickerConstants.EXTRA_MEDIA_ID, 0)
+                        viewModel.handleSelectedSiteIcon(mediaId)
+                    }
+                    data.hasExtra(MediaPickerConstants.EXTRA_MEDIA_URIS) -> {
+                        val mediaUriStringsArray = data.getStringArrayExtra(
+                                MediaPickerConstants.EXTRA_MEDIA_URIS
+                        ) ?: return
+
+                        val source = PhotoPickerMediaSource.fromString(
+                                data.getStringExtra(MediaPickerConstants.EXTRA_MEDIA_SOURCE)
+                        )
+                        val iconUrl = mediaUriStringsArray.getOrNull(0) ?: return
+                        viewModel.handleTakenSiteIcon(iconUrl, source)
+                    }
+                    else -> {
+                        AppLog.e(
+                                UTILS,
+                                "Can't resolve picked or captured image"
+                        )
+                    }
+                }
+            }
+            RequestCodes.STORIES_PHOTO_PICKER,
+            RequestCodes.PHOTO_PICKER -> if (resultCode == Activity.RESULT_OK) {
+                viewModel.handleStoriesPhotoPickerResult(data)
+            }
+            UCrop.REQUEST_CROP -> {
+                if (resultCode == UCrop.RESULT_ERROR) {
+                    AppLog.e(
+                            MAIN,
+                            "Image cropping failed!",
+                            UCrop.getError(data)
+                    )
+                }
+                viewModel.handleCropResult(UCrop.getOutput(data), resultCode == Activity.RESULT_OK)
+            }
+            RequestCodes.DOMAIN_REGISTRATION -> if (resultCode == Activity.RESULT_OK) {
+                viewModel.handleSuccessfulDomainRegistrationResult(data.getStringExtra(RESULT_REGISTERED_DOMAIN_EMAIL))
+            }
+            RequestCodes.CREATE_SITE -> {
+                viewModel.startQuickStart()
+            }
+            RequestCodes.SITE_PICKER -> {
+                if (data.getIntExtra(WPMainActivity.ARG_CREATE_SITE, 0) == RequestCodes.CREATE_SITE) {
+                    viewModel.startQuickStart()
+                }
+            }
+        }
+    }
+
+    private fun loadData(items: List<MySiteItem>) {
+        recycler_view.setVisible(true)
+        (recycler_view.adapter as? MySiteAdapter)?.loadData(items)
+    }
+
+    private fun showSnackbar(holder: SnackbarMessageHolder) {
+        snackbarSequencer.enqueue(
+                SnackbarItem(
+                        Info(
+                                view = recycler_view,
+                                textRes = holder.message,
+                                duration = Snackbar.LENGTH_LONG
+                        ),
+                        holder.buttonTitle?.let {
+                            Action(
+                                    textRes = holder.buttonTitle,
+                                    clickListener = { holder.buttonAction() }
+                            )
+                        },
+                        dismissCallback = { _, _ -> holder.onDismissAction() }
+                )
+        )
+    }
+
+    companion object {
+        private const val KEY_LIST_STATE = "key_list_state"
+        private const val KEY_NESTED_LISTS_STATES = "key_nested_lists_states"
+        private const val KEY_SITE_ID = "key_site_id"
+        fun newInstance(site: SiteModel): MySiteContentFragment {
+            val fragment = MySiteContentFragment()
+            val bundle = Bundle()
+            bundle.putSerializable(KEY_SITE_ID, site.id)
+            fragment.arguments = bundle
+            return fragment
+        }
+    }
+
+    override fun onSuccessfulInput(input: String, callbackId: Int) {
+        viewModel.onSiteNameChosen(input)
+    }
+
+    override fun onTextInputDialogDismissed(callbackId: Int) {
+        viewModel.onSiteNameChooserDismissed()
+    }
+}
